+++ packages/amino/src/index.ts
@@ -5,4 +5,4 @@ export {
   encodeBech32Pubkey,
   encodeSecp256k1Pubkey,
 } from "./encoding";
-export { PubKey, pubkeyType } from "./pubkeys";
+export { Pubkey, Secp256k1Pubkey, SinglePubkey, isSinglePubkey, pubkeyType } from "./pubkeys";