+++ packages/launchpad/src/cosmosclient.ts
@@ -1,4 +1,4 @@
-import { PubKey } from "@cosmjs/amino";
+import { Pubkey } from "@cosmjs/amino";
 import { sha256 } from "@cosmjs/crypto";
 import { fromBase64, fromHex, toHex } from "@cosmjs/encoding";
 import { Uint53 } from "@cosmjs/math";
@@ -24,7 +24,7 @@ export interface Account {
   /** Bech32 account address */
   readonly address: string;
   readonly balance: readonly Coin[];
-  readonly pubkey: PubKey | undefined;
+  readonly pubkey: Pubkey | undefined;
   readonly accountNumber: number;
   readonly sequence: number;
 }